requirements.txt
flask
flask-cors
requests
pandas
boto3
python-dotenv
openpyxl